refactor: split collection formatters
Move hash table, B-tree, and sequence payload decoding into focused
format-printer modules without changing rendering behavior.
